Finished Sprite Class Conversion

+++ LegendsReborn.Server.Base/ScriptingBase/DamageFormulaScript.cs
@@ -6,5 +6,5 @@ namespace Darkages.ScriptingBase;
 
 public abstract class DamageFormulaScript : IScriptBase
 {
-    public abstract double Calculate(Sprite obj, Sprite target, MonsterEnums type);
+    public abstract int Calculate(Sprite obj, Sprite target, MonsterEnums type);
 }
